==> chembl_32/1_act_inact_file.ipynb <==
import os
import pandas as pd
import numpy as np

code_path = os.getcwd()
project_file_path = code_path.split("chembl")[0]
training_files_path = os.path.join(project_file_path, "training_files")
training_files_path

preprocessed_tsv_file = "chembl32_preprocessed_filtered_bioactivity_dataset.tsv"

def own_read_csv(file_path, delimiter):
    """
    Reads a CSV file and returns a list of dictionaries representing the data.
    Assumes the first row contains the column headers.

    Args:
        file_path (str): Path to the CSV file.
        delimiter (str, optional): Delimiter used to separate fields in the CSV file.
            Defaults to ','.

    Returns:
        pandas.DataFrame: A DataFrame representing the data, with column headers and row values.
    """
    data = []
    with open(file_path, 'r') as file:
        lines = file.readlines()
        headers = lines[0].strip().split(delimiter)
        for line in lines[1:]:
            values = line.strip().split(delimiter)
            row_dict = dict(zip(headers, values))
            data.append(row_dict)
    return pd.DataFrame(data)

pchembl_df = own_read_csv(os.path.join(training_files_path, preprocessed_tsv_file), delimiter = "\t")
print("Len: ",len(pchembl_df))
pchembl_df.head()

# If there is no smiles value for the compound, we can not create its image so we delete it!
#versionpath = "/media/ubuntu/8TB/hayriye/DEEPScreen2.0/chembl_30"
smiles_df = pd.read_csv(training_files_path+"/chembl_32_chemreps.txt", sep = "\t")
smiles_df = smiles_df[["chembl_id","canonical_smiles"]]
smiles_df = smiles_df.rename(columns={'chembl_id': 'Compound_CHEMBL_ID'})
print("Len: ",len(smiles_df))
smiles_df.head()

pchembl_smiles_df = pd.merge(pchembl_df, smiles_df, on="Compound_CHEMBL_ID")
pchembl_smiles_df['pchembl_value'] = pchembl_smiles_df['pchembl_value'].apply(pd.to_numeric)
print("Len: ",len(pchembl_smiles_df))
pchembl_smiles_df.head()

pchembl_smiles_df.dtypes

_3cols = pchembl_smiles_df[["Target_CHEMBL_ID", "Compound_CHEMBL_ID", "pchembl_value", "canonical_smiles"]]
_3cols = _3cols.dropna()
print("Len: ",len(_3cols))
_3cols.head()

unique_targets = _3cols["Target_CHEMBL_ID"].unique()
print(len(unique_targets))
unique_targets

unique_comps = _3cols["Compound_CHEMBL_ID"].unique()
print(len(unique_comps))
unique_comps

_3cols[_3cols["Target_CHEMBL_ID"] == "CHEMBL2074"].describe()

sixty_quantile = _3cols[_3cols["Target_CHEMBL_ID"] == "CHEMBL2074"]["pchembl_value"].quantile(0.6)
sixty_quantile

float(sixty_quantile)

_3cols[ (_3cols["Target_CHEMBL_ID"] == "CHEMBL2074") & (_3cols["pchembl_value"] > sixty_quantile)] 

_3cols[ (_3cols["Target_CHEMBL_ID"] == "CHEMBL2074") & (_3cols["pchembl_value"] <= sixty_quantile)]

output_file2 = open("{}/chembl32_unique_compounds_list.txt".format(training_files_path), "w")
output_file2.write("\n".join(unique_comps) + "\n")
output_file2.close()

total_positive_count = 0
total_negative_count = 0

selected_targets = []

output_file_name = "chembl32_preprocessed_filtered_act_inact_comps_Q60.tsv"

output_file = open("{}/{}".format(training_files_path, output_file_name), "w")

df = pd.DataFrame(columns = ["Protein", "+s", "-s", "Q60"])

for i, e in enumerate(unique_targets): # 4433 targets

    protein_bioact_df = _3cols[_3cols["Target_CHEMBL_ID"] == e]
    sixty_quantile = float(protein_bioact_df["pchembl_value"].quantile(0.6))
    
    positive_bioact_df = protein_bioact_df[protein_bioact_df["pchembl_value"] > sixty_quantile]
    positive_count = len(positive_bioact_df)

    if(positive_count < 80):
        continue

    negative_bioact_df = protein_bioact_df[protein_bioact_df["pchembl_value"] <= sixty_quantile]
    negative_count = len(negative_bioact_df)

    if(negative_count < 100):
        continue

    selected_targets.append(e)
    total_positive_count += positive_count
    total_negative_count += negative_count

    print("i = {}, p = {}, + = {}, - = {}, Q60 = {:.2f}".format(i,e,positive_count, negative_count, sixty_quantile))
    df.loc[len(df)] = [e, positive_count, negative_count, round(sixty_quantile, 2)]
    
    output_file.write(e+"_act" + "\t" + ",".join(list(positive_bioact_df["Compound_CHEMBL_ID"].values)) + "\n")
    output_file.write(e+"_inact" + "\t" + ",".join(list(negative_bioact_df["Compound_CHEMBL_ID"].values)) + "\n")

print("Selected proteins = ", len(selected_targets))
print("Total positives = {}, Total negatives = {}".format(total_positive_count,total_negative_count))
output_file.close()

df.to_csv("{}/chembl32_prot_act_inact_quantile.csv".format(training_files_path), index=False)
df.head()

output_file2 = open("{}/chembl32_training_target_list.txt".format(training_files_path), "w")
output_file2.write("\n".join(selected_targets) + "\n")
output_file2.close()

# TODO Uniprot mapping 
# Create "chembl31_uniprot_mapping.txt" from "chembl31_training_target_list.txt"

total_positive_count = 0
total_negative_count = 0

selected_targets = []

path = "{}DEEPScreen2.0/training_files".format(os.getcwd().split("DEEPScreen2.0")[0])
output_file = open("{}/chembl31_preprocessed_filtered_act_inact_comps_pchembl_7.tsv".format(training_files_path), "w")

for i, e in enumerate(unique_targets):

    protein_bioact_df = _3cols[_3cols["Target_CHEMBL_ID"] == e]

    positive_bioact_df = protein_bioact_df[protein_bioact_df["pchembl_value"] > 7]
    positive_count = len(positive_bioact_df)

    if(positive_count < 80):
        continue

    negative_bioact_df = protein_bioact_df[protein_bioact_df["pchembl_value"] <= 7]
    negative_count = len(negative_bioact_df)

    if(negative_count < 100):
        continue

    selected_targets.append(e)
    total_positive_count += positive_count
    total_negative_count += negative_count

    print("i = {}, p = {}, + = {}, - = {}".format(i,e,positive_count,negative_count))
    
    output_file.write(e+"_act" + "\t" + ",".join(list(positive_bioact_df["Compound_CHEMBL_ID"].values)) + "\n")
    output_file.write(e+"_inact" + "\t" + ",".join(list(negative_bioact_df["Compound_CHEMBL_ID"].values)) + "\n")


print("Selected proteins = ", len(selected_targets))
print("Total positives = {}, Total negatives = {}".format(total_positive_count,total_negative_count))
output_file.close()
