==> chembl_32/4_create_prediction_imgs_and_json_akt1.ipynb <==
import os
import cv2
import json
#import rdkit
import numpy as np
import pandas as pd

from rdkit import Chem
from rdkit.Chem.Draw.MolDrawing import DrawingOptions
from rdkit.Chem import Draw

import cairosvg
import subprocess

def save_comp_imgs_from_smiles(tar_id, comp_id, smiles, rot=0, SIZE=200, rot_size=300):
    mol = Chem.MolFromSmiles(smiles)
    DrawingOptions.atomLabelFontSize = 55
    DrawingOptions.dotsPerAngstrom = 100
    DrawingOptions.bondLineWidth = 1.5
    # Use MolToFile(mol, path, size, imageType="png", fitImage=True)
    
    # For higher quality of image
    path_to_give_svg = os.path.join(prediction_files_path, "target_prediction_dataset", 
                                tar_id, "imgs", "{}.svg".format(comp_id))
    
    path_to_give_png = os.path.join(prediction_files_path, "target_prediction_dataset", 
                                    tar_id, "imgs", "{}.png".format(comp_id))
    
    Draw.MolToFile(mol, path_to_give_svg , size = (SIZE, SIZE ))
    cairosvg.svg2png(url = path_to_give_svg, write_to = path_to_give_png)
    subprocess.call(["rm", path_to_give_svg])
    
    # Make it larger with padding to prevent data loss while rotation
    image = cv2.imread(path_to_give_png)
    
    white_color = (255,255,255)
    full_image = np.full((rot_size, rot_size, 3), white_color, dtype = np.uint8)
    # compute center offset
    gap = rot_size - SIZE
    (cX, cY) = (gap // 2, gap // 2)
    
    # copy image into center of result image
    full_image[cY:cY + SIZE, cX:cX + SIZE] = image
    
    if rot != 0:
        # Rotate it
        (cX, cY) = (rot_size // 2, rot_size // 2)
        M = cv2.getRotationMatrix2D((cX, cY), rot, 1.0)
        full_image = cv2.warpAffine(full_image, M, (rot_size, rot_size), borderMode=cv2.INTER_LINEAR, #cv2.BORDER_CONSTANT, 
                                    borderValue = white_color)
    
    # save result
    cv2.imwrite(path_to_give_png, full_image)

protein_name = "CHEMBL4282_all"

pwd

prediction_files_path = "/home/hayriye/DEEPScreen2.0/prediction_files/"

target_prediction_dataset_path = prediction_files_path + "target_prediction_dataset/"

smiles_path = prediction_files_path + "molecule_smiles_dataset/"

smiles_file = smiles_path + "filtered_all_generated_molecules.csv"

all_df = pd.read_csv(smiles_file, sep=";")
print(len(all_df))
all_df.head()

all_df.index

all_df.iloc[0]

if not os.path.exists(target_prediction_dataset_path + protein_name + "/imgs"):
    os.makedirs(target_prediction_dataset_path + protein_name + "/imgs")

f = open(target_prediction_dataset_path + protein_name + "/prediction_dict.json", "w+")

json_dict = {"prediction": list()}
json_object = json.dumps(json_dict) 

f.write(json_object)
f.close()

total_image_count = 0
angle_list = [str(angle) for angle in range(10,360,10)]

from tqdm import tqdm

for i in tqdm(range(0, len(all_df)), desc="Predictions…", ascii=False, ncols=120):
    
    current_smiles = all_df.iloc[i]["SMILES_x"]

    current_compound_id = all_df.iloc[i]["MOL_ID"]
    
    try:
        save_comp_imgs_from_smiles(protein_name, current_compound_id, current_smiles)
        total_image_count += 1
        #print(total_image_count, current_compound_id, current_smiles)
        
        for angle in angle_list:
        
            save_comp_imgs_from_smiles(protein_name, current_compound_id+"_"+angle, current_smiles, int(angle))
            total_image_count += 1
            #print(total_image_count, current_compound_id+"_"+angle, current_smiles, int(angle))
            
    except Exception as e:
        print(e, current_compound_id, current_smiles)
    
    if i % 1000 == 0:
        print(current_compound_id, current_smiles)
        
print(total_image_count)

# removing the new line characters
with open(smiles_file) as f:
    smiles_list = [line.rstrip() for line in f]

    
print(len(smiles_list))
print(smiles_list[:10])

smiles_list[67]

compound_prefix = "GANt"
GAN_name_count = 0
total_image_count = 0
angle_list = [str(angle) for angle in range(10,360,10)]


from tqdm import tqdm

for i in tqdm(range(0, len(smiles_list)), desc="Predictions…", ascii=False, ncols=120):
    current_smiles = smiles_list[i]

#for current_smiles in smiles_list:
    
    compound_id = compound_prefix + str(GAN_name_count)
    
    try:
        save_comp_imgs_from_smiles(protein_name, compound_id, current_smiles)
        total_image_count += 1
        #print(total_image_count, compound_id, current_smiles)
        
        for angle in angle_list:
        
            save_comp_imgs_from_smiles(protein_name, compound_id+"_"+angle, current_smiles, int(angle))
            total_image_count += 1
            #print(total_image_count, compound_id+"_"+angle, current_smiles, int(angle))
            
    except Exception as e:
        print(e, GAN_name_count, compound_id, current_smiles)
    
    #if GAN_name_count % 1000 == 0:
        #print(GAN_name_count, compound_id, current_smiles)
    GAN_name_count += 1
